allow stop tasks to have names and tags
For longer running tasks, there may only be an end to bottom most task,
making it unclear what was the original unknown start.  This change
makes it so the stop can have an optional task name, and tag id.

Task start should have precedence over stop for task name and tag id.
If start and stop task names are non-null, then they should match, or
else it would be an obvious bug.  Not sure what to do with it.