Add a JCStress test
After trying this out on my x86_64 processor with different buffer
sizes (16, 256, 8192, 65536) and word types (byte, int) I was not able
to get this to race and end up with a corruption.  Having the `+1` in
the end index for the reader is noticeable, causing the race failures to
occur.  Smaller buffers are easier to trigger the race with.

The code is from a copy of the JCStress sample maven project, so the
plugins to generate the jar don't work with gradle.

Note that each pair of actors operates on a new object, so the shared
state for this has to be in static variables.  Otherwise the state is
reset for each pair of actors.