api: remove CompileTimeConstant
The original reason for this was to mirror the predecessor to PerfMark.  However,
the reason for this was to aid in the agent to rewrite the API calls, which
PerfMark doesn't depend on.

While it is good for these values to be constant, the annotation is not capable of
expressing other, useful types of constants (like Class.getName), and isn't
generally enforceable.  Instead, let this value be non-compile time constant,
but instead "runtime constant".  This term is intentionally vague, but roughly it
means that the cardinality of task and event names should scale with respect to the
amount of code, rather than the input size of the program.