Exclude everything in .gradle/cache but modules-2
The other folders have contents that may change every run, and which ones
change is dependent on the gradle version. The other folders also don't seem to
have much that's useful. The plugin-resolution removal is also then doubly
unnecessary, as it is 1) not in modules-2 and 2) I don't see it being created
any more in Gradle 5.x.

Previously it was mirroring the Gradle documentation[1], so it is now clear the
docs are out-of-date.

1. https://docs.travis-ci.com/user/languages/java/#caching