api,impl: add a delayed tag impl
This allows delaying a tag being attached until later, in case the tag processing is expensive. I only added it for tags, but a similar pattern can be used for events and tasks. I gave myself some leeway with the implementation, but decided to be conservative with this PR. Rather than stringifying the args on /reading/, it stringifies them on write. I can't think of a clear way of expressing this distinction in the method name; ideas are welcome. Stringing at write incurs a possible waste of work, but avoids synchronization and memory retention issues. Stringing at read reduces the overhead of adding tags.