all: Add automatic module support
This makes Perfmark compatible with thew Java 9 Module system. For classes compiled for 8 and lower, an Automatic-Module-Name field is added to the MANIFEST.MF, to ensure a stable module name. (and not derive it from the jar name). For the java9 package, an actual module descriptor is added, which I have tested locally works.