impl,agent: Remove Markers and replace it with String calls
While less precise, the string version of callsites is much easier to work with. It avoids needing to mess with clinit in classes, avoids needing to check that Marker is on the classpath, and avoids a 2 pass scan of the class file. Marker is nice, but adding the call site via a tag is much simpler to implement.

This also removes the Java formatter plugin, as it doesn't yet work JDK16, and appears to be unmaintained.