api: avoid logging in static init.
The logging system in Java is pretty heavy weight, causing a lot of class loading to happen, Service Loaders, and Shutdown hooks.   It is expected that PerfMark is included in a lot
of places, but only occasionally accompanied by the implementation.    Avoid logging in the cases where the impl is absent. These were already FINE level logs usually, so they would
have been invisible in the common case.

Additionally, this also tries to avoid throwing an exception in the event a security manager is present.  Both Logger init and Boolean.getBoolean have Security Manager implications.
While the Secutiy is slated to go away in a later version of Java, it is still present in the versions PerfMark targets.